Implement msg_unlisten.
It does *not* check that all associated connections are closed. For
example, in udp, there might be a number of application-level udp
connections active and depending on the socket being closed here. It's
up to the user to close them, but I'm interested in adding a check for
this in DEBUG mode.