Clarify which fn implementations are for which os.
Before doing this, a coder could search for a function implementation,
and it wouldn't have been obvious if they were looking at the mac/linux
version or the windows version of the code. Now this is clarified at the
top of each function. (This also indicates which functions are twice
implemented.)