Rename src/ -> msgbox/
I'm doing this to make it easier to include the library in projects.
When I add third-party tools, I often like to copy over the source files
themselves. This is made easier if all I have to do is copy a directory.
For example, if the docs suggest this as an official way to use a
library, then I have the peace of mind of knowing that I haven't
accidentally under- or over-copied over the source files.

This also drops the need for library-users to make an arbitrary decision
of what to call the directory they put these files into.

I can standardize my example code to include "msgbox/msgbox.h", and this
pattern is one that makes sense to me as a general way for
library-writers to support inclusion.