Fix #4: a msg_send off a reply is not a reply.
In a few more words for added clarity: It used to be the case that if
you called msg_send from your msgbox-event-handler after you were
triggered by a reply to a request you had previously sent, then that
msg_send would itself look like a reply to the other party. This was a
bug. It is now squashed.