Fix tcp timeout messages; no context on msg_listen
What I mean by "no context on msg_listen" is that I removed the
conn_context parameter to that function and deprecated the idea that a
connection has a context without a specific remote address associated
with it. Specifically, listening connections don't support a persistent
context outside the context of interactions with a specific endpoint.

In slightly more detail: tcp connections are always dedicated because we
get a new socket and a new connection object with each machine/machine
pair. udp is trickier in that the server only ever has a single socket
for all clients, and that socket comes with a single msg_Conn object. We
effectively provide multiple conn_context values to the user using
ConnStatus objects behind the scenes.

I could have opted to persist a kind of meta-context for listening
connections, but I decided not to as all my use cases involve a small
number of listeners where it's practical to treat a context specially,
such as by use of file-scope variables (static globals).