Fix #7.
The fix works by wrapping user callbacks so that the msg_Conn they see
has conn->conn_context both loaded before the call, and saved
afterwards. The persistant location for a udp connection's conn_context
is the associated ConnStatus struct.

In the future, I'd be interested in directly exposing the memory
location of the conn_context to the user, so that there's less
cleverness involved. Even though that memory is managed by msgbox, we
could clearly indicate when it becomes obsolete through an event to the
user.

For now, I think this solution is sufficient. Basically, the
conn_context will behave sanely in all the cases I an currently think of
:)