Add a new test and a fix for it.
Basically, the way conn_context data was saving and restoring was buggy.
Essentially, the current setup can only handle user writes to
conn_context when they happen in a callback (from msgbox to user code).
This is because the user may change the remote address stored in a
msg_Conn struct without changing the conn_context, and it's not clear
what we should do in that situation.

The solution that makes the most sense to me (which is implemented here)
is to let the user change the remote address of a listening udp
connection whenever they want so they can send things anywhere at any
time, but only save conn_context data within callbacks, and that context
is always associated with the remote address that *initiated* the given
callback, as opposed to whatever remote address the user may have set in
the meantime.

To be honest, all of this design is sub-ideal. Really, there should be
one struct per conceptual connection, and these subtleties would become
much much easier to deal with. Alas, that improved design is a chunk of
work that I defer for now.